git_fix_3.0

+++ tasks/seq/naumov_b_min_colum_matrix/perf_tests/main.cpp
@@ -79,5 +79,5 @@ TEST(sequential_naumov_b_min_colum_matrix_perf_test, test_task_run) {
 
   ppc::core::Perf::print_perf_statistic(perfResults);
 
-  ASSERT_EQ(out.size(), cols);
+  ASSERT_EQ(out.size(), static_cast<size_t>(cols));
 }
